Lab Assignment 10
Assignment Ten – Working with extended libraries in p5.js

This assignment involves me experimenting with the p5.Polar library. This library allows you to create radial patterns, like a kaleidoscope, which I utilized to create a flower. I used the main function the library offers, setcenter, to give the code an exact point of where the center of my code is. After that, I used both polarEllipses and polarTriangle to create the shapes in a radial formation to depict my blue flower.